requirements.txt
Flask
transformers
torch
torchvision
torchaudio
requests
python-dotenv